[SDR-410] refactor: 식당 목록, 유저 리뷰 탭 컴포넌트 생성 후 로직 분리
