+++ fe/.eslintrc
@@ -21,6 +21,7 @@
     "import/order": 0,
     "prefer-object-spread": 0,
     "react-hooks/exhaustive-deps": 0,
-    "@typescript-eslint/no-shadow": 0
+    "@typescript-eslint/no-shadow": 0,
+    "react/no-this-in-sfc": 0
   }
 }
